improve models

+++ aquaholic/models.py
@@ -15,13 +15,13 @@ class UserInfo(models.Model):
     UserInfo class for collect user information.
     """
     user = models.ForeignKey(User, on_delete=models.CASCADE, null=True)
-    weight = models.DecimalField(max_digits=10, decimal_places=2, default=0)
-    exercise_time = models.DecimalField(max_digits=10, decimal_places=2, default=0)
-    water_amount_per_day = models.DecimalField(max_digits=10, decimal_places=2, default=0)
+    weight = models.FloatField(default=0)
+    exercise_time = models.FloatField(default=0)
+    water_amount_per_day = models.FloatField(default=0)
     first_notification_time = models.TimeField('first notification time', null=True)
     last_notification_time = models.TimeField('last notification time', null=True)
-    total_hours = models.DecimalField(max_digits=10, decimal_places=10, null=True)
-    water_amount_per_hour = models.DecimalField(max_digits=10, decimal_places=2, null=True)
+    total_hours = models.FloatField(null=True)
+    water_amount_per_hour = models.FloatField(null=True)
     notify_token = models.CharField(max_length=200, null=True)
 
     def get_water_amount_per_day(self):
@@ -49,7 +49,7 @@ class Schedule(models.Model):
     """Schedule class for create notification time."""
     user_info = models.ForeignKey(UserInfo, on_delete=models.CASCADE, null=True)
     notification_time = models.DateTimeField('notification time', null=True)
-    expected_amount = models.DecimalField(max_digits=10, decimal_places=10, default=0)
+    expected_amount = models.FloatField(default=0)
     notification_status = models.BooleanField(default=True)
 
     def change_notification_status(self):
@@ -63,5 +63,5 @@ class Schedule(models.Model):
 class Intake(models.Model):
     """Intake class for collect water intake of user per day."""
     user_info = models.ForeignKey(UserInfo, on_delete=models.CASCADE, null=True)
-    user_drinks_amount = models.DecimalField(max_digits=10, decimal_places=10, default=0)
+    user_drinks_amount = models.FloatField(default=0)
     intake_date = models.DateTimeField(default=timezone.now)
